+++ CHANGELOG.md
@@ -10,7 +10,10 @@
     **Vol/Vol.**, **Volume**, **Pt** too. Previously we only considered **Part**.
     - **Compilation - Volume 2**
     - Compilation Volume 2 -> **Compilation, Volume 2**
-    - Vol 2 - Compilation -> **Compilation, Vol 2**
+    - If series is in the beginning of the album, it is moved to the end
+      - Vol 2 - Compilation -> **Compilation, Vol 2**
+    - For **Vol** we also ensure the space and removal of leading zeroes
+      - Vol.02 -> **Vol. 2**
 
 * `albumtype`: the EP albumtype is recognized for split EPs.